requirements.txt
beautifulsoup4
pdfplumber
streamlit==1.35.0
langchain==0.2.14
langchain_groq==0.1.9
pandas==2.0.2
requests
python-dotenv
groq